(extension) flatten danmaku source options
