+++ packages/danmaku-anywhere/src/common/localization/en/translation.ts
@@ -278,6 +278,13 @@ const pages = {
             'Protobuf：Slower but more danmaku. Up to 1000 danmaku per minute. Slower for longer videos.',
         },
       },
+      tencent: {
+        limitPerMin: 'Rate Limit (per minute)',
+        help: {
+          limitPerMin:
+            'Maximum number of danmaku requests per minute to avoid rate limiting.',
+        },
+      },
     },
     retentionPolicy: {
       enabled: 'Enable Retention Policy',
