+++ packages/danmaku-anywhere/src/popup/pages/options/pages/advanced/components/AnalyticsOption.tsx
@@ -12,9 +12,8 @@ export const AnalyticsOption = () => {
 
   return (
     <ToggleListItemButton
-      enabled={data.debug}
+      enabled={data.enableAnalytics}
       onToggle={handleToggle}
-      onClick={handleToggle}
       itemText={t('optionsPage.enableAnalytics')}
       isLoading={isLoading}
     />
