+++ packages/danmaku-anywhere/src/popup/pages/options/pages/advanced/components/DebugOption.tsx
@@ -12,7 +12,6 @@ export const DebugOption = () => {
     <ToggleListItemButton
       enabled={data.debug}
       onToggle={handleToggle}
-      onClick={handleToggle}
       itemText="Debug"
       isLoading={isLoading}
     />
