+++ packages/danmaku-anywhere/src/popup/pages/options/pages/danmakuSource/components/DanDanPlayOptions.tsx
@@ -1,5 +1,5 @@
 import { zodResolver } from '@hookform/resolvers/zod'
-import { Box, MenuItem, Stack, TextField } from '@mui/material'
+import { MenuItem, TextField } from '@mui/material'
 import { useMutation } from '@tanstack/react-query'
 import { produce } from 'immer'
 import { useEffect } from 'react'
@@ -7,16 +7,10 @@ import { Controller, useForm } from 'react-hook-form'
 import { useTranslation } from 'react-i18next'
 
 import { useToast } from '@/common/components/Toast/toastStore'
-import {
-  DanmakuSourceType,
-  localizedDanmakuSourceType,
-} from '@/common/danmaku/enums'
 import { ChConvertList } from '@/common/options/extensionOptions/constant'
 import type { DanmakuSources } from '@/common/options/extensionOptions/schema'
 import { danmakuSourcesSchema } from '@/common/options/extensionOptions/schema'
 import { useExtensionOptions } from '@/common/options/extensionOptions/useExtensionOptions'
-import { OptionsPageToolBar } from '@/popup/component/OptionsPageToolbar'
-import { OptionsPageLayout } from '@/popup/layout/OptionsPageLayout'
 
 export const DanDanPlayOptions = () => {
   const { t } = useTranslation()
@@ -47,7 +41,7 @@ export const DanDanPlayOptions = () => {
       )
     },
     onSuccess: () => {
-      toast.success(t('common.success'))
+      toast.success(t('common.saved'))
     },
   })
 
@@ -62,36 +56,27 @@ export const DanDanPlayOptions = () => {
   }, [formData])
 
   return (
-    <OptionsPageLayout>
-      <OptionsPageToolBar
-        title={t(localizedDanmakuSourceType(DanmakuSourceType.DanDanPlay))}
-      />
-      <Box p={2}>
-        <Stack mt={2} gap={2}>
-          <Controller
-            name="dandanplay.chConvert"
-            control={control}
-            render={({ field }) => {
-              return (
-                <TextField
-                  {...field}
-                  select
-                  fullWidth
-                  label={t('optionsPage.chConvert.name')}
-                  error={!!errors.dandanplay?.chConvert}
-                  helperText={errors.dandanplay?.chConvert?.message}
-                >
-                  {ChConvertList.map((option) => (
-                    <MenuItem value={option.value} key={option.label}>
-                      {t(option.label)}
-                    </MenuItem>
-                  ))}
-                </TextField>
-              )
-            }}
-          />
-        </Stack>
-      </Box>
-    </OptionsPageLayout>
+    <Controller
+      name="dandanplay.chConvert"
+      control={control}
+      render={({ field }) => {
+        return (
+          <TextField
+            {...field}
+            select
+            fullWidth
+            label={t('optionsPage.chConvert.name')}
+            error={!!errors.dandanplay?.chConvert}
+            helperText={errors.dandanplay?.chConvert?.message}
+          >
+            {ChConvertList.map((option) => (
+              <MenuItem value={option.value} key={option.label}>
+                {t(option.label)}
+              </MenuItem>
+            ))}
+          </TextField>
+        )
+      }}
+    />
   )
 }
