+++ packages/danmaku-anywhere/src/popup/router/router.tsx
@@ -17,8 +17,6 @@ import { IntegrationPolicyEditor } from '../pages/integrationPolicy/pages/Integr
 import { MountPage } from '../pages/mount/MountPage'
 import { Options } from '../pages/options/Options'
 import { DanmakuSource } from '../pages/options/pages/danmakuSource/DanmakuSource'
-import { BilibiliOptions } from '../pages/options/pages/danmakuSource/pages/BilibiliOptions'
-import { DanDanPlayOptions } from '../pages/options/pages/danmakuSource/pages/DanDanPlayOptions'
 import { HotkeyOptions } from '../pages/options/pages/hotkeyOptions/HotkeyOptions'
 import { RetentionPolicyPage } from '../pages/options/pages/retentionPolicy/RetentionPolicyPage'
 
@@ -118,16 +116,6 @@ export const router: ReturnType<typeof createHashRouter> = createHashRouter([
       {
         path: 'danmaku-source',
         Component: DanmakuSource,
-        children: [
-          {
-            path: 'dandanplay',
-            Component: DanDanPlayOptions,
-          },
-          {
-            path: 'bilibili',
-            Component: BilibiliOptions,
-          },
-        ],
       },
 
       {
